Successful Pathfinding with jumps
First successful implementation of platformer pathfinding with jumps. To achieve this the Node script has been modified to hold information whether a jump is required to reach the next node or not. This now is determined during FindPath method when it goes through each walk neighbour or jump neighbour. To retain this information when the constructing a path, RetracePath method and FollowPath methods have been changed to use an array of Nodes for a path as opposed to simply using Vector3s for a path.